Relocating the model interfaces to the clients.
