--- src/events.ts
@@ -1,78 +1,0 @@
-import {
-    Battery,
-    Calendar,
-    Headlight,
-    MowerMetadata,
-    MowerState,
-    Planner,
-    Position
-} from './model';
-
-/**
- * Describes a connected event.
- */
-export interface ConnectedEvent {
-    connectionId: string;
-    connected: boolean;
-}
-
-/**
- * Describes an automower event.
- */
-export interface AutomowerEvent {
-    id: string;
-    type: AutomowerEventTypes;
-}
-
-/**
- * Describes the automower event types.
- */
-export enum AutomowerEventTypes {
-    UNKNOWN = 'unknown-event',
-    STATUS = 'status-event',
-    POSITIONS = 'positions-event',
-    SETTINGS = 'settings-event'
-}
-
-/**
- * Describes a status event.
- */
-export interface StatusEvent extends AutomowerEvent {
-    attributes: {
-        battery: Battery;
-        mower: MowerState;
-        planner: Planner;
-        metadata: MowerMetadata;
-    };
-}
-
-/**
- * Describes a positions event.
- */
-export interface PositionsEvent extends AutomowerEvent {
-    attributes: {
-        positions: Position[];
-    };
-}
-
-/**
- * Describes a settings event.
- * <p>
- * This event occurs when the settings have been modified on a mower.
- */
-export interface SettingsEvent extends AutomowerEvent {
-    attributes: {
-        calendar?: Calendar;
-        cuttingHeight?: number;
-        headlight?: Headlight;
-    };
-}
-
-/**
- * Describes an error event.
- */
-export interface ErrorEvent {
-    error: string;
-    message: string;
-    type: string;
-}
