+++ src/services/policies/mowerInMotionPolicy.ts
@@ -1,4 +1,4 @@
-import { Activity, MowerState } from '../../model';
+import { Activity, MowerState } from '../../clients/automowerClient';
 import { Policy } from './policy';
 
 /**
