+++ test/clients/automowerEventStreamClient.spec.ts
@@ -2,9 +2,9 @@ import { It, Mock, Times } from 'moq.ts';
 
 import * as constants from '../../src/settings';
 
+import { AutomowerEvent, AutomowerEventTypes, ConnectedEvent, ErrorEvent } from '../../src/clients/automowerEventStreamClient';
 import { WebSocketWrapper } from '../../src/clients/primitives/webSocketWrapper';
 import { PlatformLogger } from '../../src/diagnostics/platformLogger';
-import { AutomowerEvent, AutomowerEventTypes, ConnectedEvent, ErrorEvent } from '../../src/events';
 import { AutomowerEventStreamClientImplSpy } from './automowerEventStreamClientImplSpy';
 
 describe('AutomowerEventStreamClientImpl', () => {
