+++ test/services/homebridge/abstractContactSensor.spec.ts
@@ -3,8 +3,8 @@ import { API, HAP, PlatformAccessory } from 'homebridge';
 
 import { It, Mock, Times } from 'moq.ts';
 import { AutomowerContext } from '../../../src/automowerAccessory';
+import { MowerMetadata } from '../../../src/clients/automowerClient';
 import { PlatformLogger } from '../../../src/diagnostics/platformLogger';
-import { MowerMetadata } from '../../../src/model';
 import { Policy } from '../../../src/services/policies/policy';
 import { ContactSensorSpy } from './contactSensorSpy';
