+++ test/services/husqvarna/automower/discoveryService.spec.ts
@@ -3,8 +3,8 @@ import { It, Mock, Times } from 'moq.ts';
 import { AutomowerAccessory } from '../../../../src/automowerAccessory';
 import { AutomowerAccessoryFactory } from '../../../../src/automowerAccessoryFactory';
 import { AutomowerPlatform } from '../../../../src/automowerPlatform';
+import { Activity, HeadlightMode, Mode, Mower, OverrideAction, RestrictedReason, State } from '../../../../src/clients/automowerClient';
 import { PlatformLogger } from '../../../../src/diagnostics/platformLogger';
-import { Activity, HeadlightMode, Mode, Mower, OverrideAction, RestrictedReason, State } from '../../../../src/model';
 import { DiscoveryServiceImpl } from '../../../../src/services/husqvarna/automower/discoveryService';
 import { GetMowersService } from '../../../../src/services/husqvarna/automower/getMowersService';
