+++ test/services/husqvarna/automower/getMowersServiceImpl.spec.ts
@@ -1,8 +1,8 @@
 import { Mock, Times } from 'moq.ts';
 
-import { AutomowerClient } from '../../../../src/clients/automowerClient';
+import { Activity, AutomowerClient, HeadlightMode, Mode, Mower, OverrideAction, RestrictedReason, State } from '../../../../src/clients/automowerClient';
 import { NotAuthorizedError } from '../../../../src/errors/notAuthorizedError';
-import { AccessToken, Activity, HeadlightMode, Mode, Mower, OverrideAction, RestrictedReason, State } from '../../../../src/model';
+import { AccessToken } from '../../../../src/model';
 import { AccessTokenManager } from '../../../../src/services/husqvarna/accessTokenManager';
 import { GetMowersServiceImpl } from '../../../../src/services/husqvarna/automower/getMowersService';
