+++ test/services/pauseSwitch.spec.ts
@@ -3,8 +3,8 @@ import { API, CharacteristicEventTypes, CharacteristicSetCallback, Characteristi
 import { It, Mock, Times } from 'moq.ts';
 
 import { AutomowerContext } from '../../src/automowerAccessory';
+import { Activity, Mode, MowerMetadata, MowerState, State } from '../../src/clients/automowerClient';
 import { PlatformLogger } from '../../src/diagnostics/platformLogger';
-import { Activity, Mode, MowerMetadata, MowerState, State } from '../../src/model';
 import { NameMode } from '../../src/services/homebridge/abstractSwitch';
 import { MowerControlService } from '../../src/services/husqvarna/automower/mowerControlService';
 import { MowerIsPausedPolicy } from '../../src/services/policies/mowerIsPausedPolicy';
