+++ test/services/policies/mowerIsArrivingPolicy.spec.ts
@@ -1,4 +1,4 @@
-import { Activity, Mode, State } from '../../../src/model';
+import { Activity, Mode, State } from '../../../src/clients/automowerClient';
 import { DeterministicMowerIsArrivingPolicy } from '../../../src/services/policies/mowerIsArrivingPolicy';
 
 describe('DeterministicMowerIsArrivingPolicy', () => {
