+++ test/services/scheduleSwitch.spec.ts
@@ -3,8 +3,8 @@ import { API, HAP, PlatformAccessory } from 'homebridge';
 import { It, Mock, Times } from 'moq.ts';
 
 import { AutomowerContext } from '../../src/automowerAccessory';
+import { Activity, Calendar, Mode, MowerMetadata, MowerState, Planner, RestrictedReason, State } from '../../src/clients/automowerClient';
 import { PlatformLogger } from '../../src/diagnostics/platformLogger';
-import { Activity, Calendar, Mode, MowerMetadata, MowerState, Planner, RestrictedReason, State } from '../../src/model';
 import { NameMode } from '../../src/services/homebridge/abstractSwitch';
 import { DISPLAY_NAME } from '../../src/services/homebridge/characteristics/cuttingHeight';
 import { ChangeSettingsService } from '../../src/services/husqvarna/automower/changeSettingsService';
